refactor(io): extract GFF3 CDS parser to treetime-io
Move GFF3 parsing from `commands/ancestral/aa_node_data.rs` to `treetime-io/src/gff.rs` so the parser is discoverable and reusable by future callers (GenBank support, mugration annotations). The command module retains only the augur-specific conversion from `GffCdsFeature` to `AugurNodeDataJsonAnnotationEntry`.

- Add `treetime-io::gff` module with `GffCdsFeature`, `GffCdsSegment`, `read_gff3_cds_features`, `read_gff3_cds_features_filtered`, `resolve_cds_name`, `parse_gff_attributes`
- Replace inline parsing in `aa_node_data.rs` with `read_gff3_cds_features_filtered` + `gff_cds_to_annotation` conversion
- Move all GFF3 tests (basic, minus-strand, non-multiple-of-3, ##FASTA, seqid, name priority) to `treetime-io`
