+++ packages/treetime-io/src/lib.rs
@@ -7,6 +7,7 @@ pub mod csv;
 pub mod dates_csv;
 pub mod discrete_states_csv;
 pub mod fasta;
+pub mod gff;
 pub mod graph;
 pub mod graphviz;
 pub mod nex;
